Pose Sketching: First working prototype of "Direct Mode"
This commit introduces a new tool which allows users to draw the shape that
a chain of bones should form, and have the bones be transformed to fit.

Usage:
1) Select a chain of bones. You must select at least 2 bones, and they should
   ideally be connected (or appear to be) for best results.
2) EKEY to invoke the operator
3) Draw a stroke  (*)
4) Your rig should now be posed

(*) This operator uses Grease Pencil strokes to provide its sketch input.
    Please ensure that strokes are drawn in 3D space (not view), and that
    "Continous Drawing" is not enabled.

The currently implementation is still heavily WIP, and suffers from a number
of limitations. It is also just "simpler" of the two initial algorithms planned.
* It currently only sets the pose_mat values (and the pose_head/tail) values.
  To keyframe any poses made this way, use "Visual Keying" or else the poses
  won't hold.

* There is currently a bug where each subsequent attempt to pose the rig
  will result in incorrect bone alignments. More work is required to figure
  out why this occurs.