PSculpt: Experimental "Draw" brush implementation hacked together from the Trackball transform code
This is still very much an initial, hacked-together implementation that lacks
a lot of polish. It is based on the existing code for the "Trackball" rotation
tool. I've just taken that code (specifically the parts related to bones),
and hacked everything together so that it can be used via Pose Sculpting brushes.

Usage:
Currently, it seems to work best if you do a "dab and drag" action, where you
target a particular bone (or cluster of bones) and make a small movement to
adjust the way the bone is orientated.

Todo's:
* Hook up the strength settings
* Hook up channel locking magic (and/or caching of original values)
* Play with a hybrid Grab + Trackball brush
* Review the selection of the transform centerpoint

* Remove the old and defective "Draw" brush code (now named "comb"). This new
  brush type may still have to be renamed to "Trackball" or something similar
  if we come up with something better.