PSculpt: Rename Draw/Trackball -> Adjust
After a few days of playing around with this, it seems that, while close to the
behaviour we want, "trackball" doesn't quite act as a general purpose "Draw"
(i.e. you'd expect to be able to do line of action stuff with that, and trackball
fails for that case). However, trackball excels at another situation: tweaking
and adjusting existing poses to add a bit more character, or to generally improve the
pose "in place". Hence the decision to retain and rename this brush.

Notes:
* For now, "Draw" still does the same thing as Adjust. It'll stay that way until
  I come up with a better draw brush
* Knocked out "Radial" and "Wrap" from the types enum for now. They'll come back
  later when we actually implement them. Until then, they'll stay hidden, in case
  we just want to knock them out completely (and to avoid confusing testers)