(Un-)Register widgets attached to modal operator with the operatortype
We have some widgets-groups that are attached to modal operators, meaning, they are only visible (actually accessible) as long as the modal OP runs. I simplified the process of setting those up by adding a wmWidgetGroupType pointer to wmOperatorType struct. If you assign a value to it, everything else is handled by widget system.

Also, previously key map entries for widget-groups attached to modal OP's were only visible while the OP ran, since it was registered and unregistered during OP runtime. This is fixed, since they are now (un-)registered with the OP-type.