Make key maps for widgets visible in key map editor
For each editor that has widgets, we draw a "Widgets" category into it's key map list. Under this we have the key maps for the individual widget group types.

Having to hardcode the entries in keyconfig_utils.py is not so nice, but don't know of a better way to do this.