Cycles: Experiment with using Lambert cylindrical equal-area projection
This commit implements new panorama and environment texture projection
which is called after a dude named Lambert:

  https://en.wikipedia.org/wiki/Lambert_cylindrical_equal-area_projection

Advantage of this projection is that it doesn't waste sampling time next
to the poles (which is a huge issue for the equirectangular projection).

Disadvantage of this model is that there'll be some distortion on the
poles when mapping it back to the sphere.

This is partially worked around by using scaling magic so there'll be
a bit more samples added on the top/bottom of the image (which makes it
technically a Roosendaal-Sharybin mapping since that's what we've been
discussing in the studio).

Maybe we should be more aggressive here with clamping of the sine, but
let's start doing some real-world tests first.