HMD: Add support of lazil loading udev library
This way we can statically link against hidapi libraries but dont'
require static linking against udev or even having specific ABI
version of this library installed on the client machine.

For now the following libraries will try to be loaded:
  libudev.so, libudev.so.0, libudev.so.1 and libudev.so.2

Not future proof since it might be version 3 of udev releases
eventually, good enough for the beginning. We can simply do a
for-loop to query like 100 of ABIs.

Udev wrangler is a bit different from other wranglers we've got,
it has wrapper functions around dlopen-ed symbols, so you can't
check directly if function was loaded or not, but this is the
only way to make external libraries working fine with wrangler
without need of re-compiling the library.