Cycles: Commit hack which allows to render less samples on the poles
This is more an experiment for now to see how it could improve render
times without loosing too much quality. Implementation is quite hacky
due to timing constraints here.

Now it is possible to control how much samples will be rendered on
the pole, so now regular Samples are controlling samples in the middle
of the frame, and those samples are falling off to the pole samples
using sine interpolation.

Limitations:

- Only works for final renders without progressive refine option.
- It is per-tile samples, so tile size should be small enough to
  see an effect.

This is not something even nearly considered for master, more like
a research and test project.