Cycles: Use dedicated BVH for triangles and hair
This is a required step forward OBB type of BVH which will be more optimal
for hair but less optimal for triangles. Which means, we'll need to have
dedicated hair BVH once we'll start working on OBB BVH.

This commits does exactly that: there are now two separate BVH, one is
fully dedicated to triangles and another one is fully dedicated to hair.
They are built separately and also traversed separately.

This allows:

- Get rid of triangle storage for curve primitives, which solves memory
  usage issue without slowdown of triangles intersection test.

- Perform test against triangles for shadow rays, which will often lead
  to som speed gain because ray often hits opaque triangle and in this
  case it's not needed to do any ray-to-curve intersections.

Unfortunately, current implementation used 7 more textures, which is quite
bad on lower end GPUs due to tight number of textures limit, which means
less image textures can be used now.

Here are some numbers:

                               Master          Patch
Koro.blend
  Memory during render(Mb):    910             814
  Peak memory(Mb):             984             963
  Render time:                 5m 51s          5m 21s

Bunny.blend
  Memory during render(Mb):    8034            6375
  Peak memory(Mb):             9058            7713
  Render time:                 9, 25s          7m 44s

There are some render result difference with low number of samples,
probably because of different order of geometry hits when there are
multiple primitives at the same depth. But that's to be doublechecked.

There should be no affect on scenes with just triangles.

Since there are some interesting results already decided to put it
online for some more tests.