Cycles: Initial implementation of unaligned BVH builder
The idea is to use unaligned boundbox nodes for hair BVH, which should
in theory improve performance of ray-to-hair intersection.

Only tree building is done currently, still needed to work on packing
and traversal code.

The idea oh builder is the following:

- We calculate SAH of regular split
- Then we calculate alignment orientation for the BVH range, using
  first curve from the range to define orientation.
- Then we calculate SAH using alignment from the previous step.
- If SAH calculated for a found orientation is better than regular
  orientation we create unaligned node.
- All leaf nodes are unaligned.

There is some code duplication which we can avoid, and some parts are
not really optimized yet.