Cycles: SIMD optimization of unaligned BVH traversal
Some movement towards SIMD-optimized traversal of unaligned nodes,
not fully optimal yet, but now the branch is about 4% faster on
a test scene with hairy monkey with 40K hair.

Should still be possible to optimize matrix multiplication and
maybe some parts could be re-formulated a bit to become more
friendly for vectorization.

But likely next step would be to support QBVH for hair, that should
be easier to vectorze with similar to current formulation.