Category: Improved Bezier Curve Extrusion
This feature allows the user to have more control over the extrusion of Bezier
curves. This is done by making possible unidirectional extrusion.

On a code level, the curve.c file in blenkernel was modified to take into
account these options while generating the extruded object.

The feature stops working with non-zero bevel.