Improved Bezier curve extrusion
This patch attempts to solve the issues with D1841.

Category: Improved Bezier Curve Extrusion

This feature allows the user to have more control over the extrusion of Bezier
curves. This is done by making possible unidirectional extrusion.

On a code level, the curve.c file in blenkernel was modified to take into
account these options while generating the extruded object.

The feature stops working with non-zero bevel.

Merge remote-tracking branch 'origin/master' into gsoc2016-improved_extrusion

Category: Improved Bezier curve extrusion with non-zero bevel

On a previous commit, the Bezier curve extrusion was improved. However, when
non-zero bevel values were set, the new extrusion options stopped having effect.
This commit fixes that.

On a code level, the code used for extrusion and for bevel is a duplicate.
On the first commit only the extrusion code was modified. On this commit,
the bevel code was modified in the same way, to take into account the improved
extrusion options.

Differential Revision: https://developer.blender.org/D2017