Add "Layer Manager" editor
We didn't decide yet if the layer manager is going to be placed in an own editor, in the outliner or anywhere else. For now, I'll use this new editor since it allows me to focus on creating a really rough the layer manager GUI. Later we can merge into an other editor if needed.