Depsgraph: Fallback to plain array for operations once depsgraph was fully built
Storing dozen operations in a map is not really efficient for traversal and
looping (which happens A LOT) during graph evaluation. So let's sue map during
building process to speed up some lookups but once graph was finalized we are
now switching to an array for operations.

This solves the speed regression caused by switch to a native BLI utilities
and also gives few percent extra speedup comparing to master. So it's a win.