Support assigning objects to layers
Selected objects are now assigned to selected layers by pressing the M-key in the layer manager editor. Nothing visible happens then, but objects are now assigned to the layer.
Deleting a layer now invokes a popup from which the user can choose to select the layer with or without content.

Code still needs more work, but this depends on other changes. Undo/redo and file read/write doesnt' work yet, it's simply crashing.