Viewport: Draw objects by iterating over layers
Instead of iterating over all objects, we now iterate over layers and the objects inside of them. Later we can skip invisible layers, but they are not supported yet.
Objects assigned to multiple layers are currently drawn multiple times, we can avoid this (although it *might* be useful in some cases).

Since current code allows objects not being assigned to any layer, un-assigned objects are not drawn at all. To fix this you need to assign them to a layer.