Use a hash map to store object reader / object path pair to be able to reconstruct the object hierarchy more easily.
The previous approach was brute forcing by iterating over all the
objects in the scene for each object creating to find the object
(parent) that has the same name as its parent in the abc archive.

This was both inefficient and prone to errors since multiple objects in
the abc archive can have the same name, as long as their path are
different, but Blender only allows for unique object name, so the wrong
object could have been choosen as the parent of a given object.

Now all of the tests file import correctly.