Revert geometry cache
Reverting the geometry cache for now, there's still a lot of work to be done on
the other parts of the subdivision code and the added complexity of the cache
is slowing down development. Will add it back later after things are more
functional. For now this means scenes will potentially use much more memory, so
care should be taken when using the feature.