Store array of all layer items in the layer tree
Coexists with LayerTree.items and LayerTreeItem.childs listbases which were used for recursive iterations over the layer tree (using BKE_layertree_iterate or BKE_layeritem_iterate_childs). However, it was a bit annoying having to use this everywhere, especially if many args had to be passed as void *. Also, iterative methods with arrays instead of listbases should give us a bit of speedup.
It's possible LayerTree.items will not be needed anymore, but will leave in for now.

Array is only used on BKE level so far.