Pack patch info into triangle data
Pack data for finding the originating patch of a triangle into unused bytes
in __tri_verts, __tri_vnormal and __tri_vindex. Since these bytes weren't
being used there's no increase in memory usage and we avoid more complicated
logic that would be necessary if this data was stored in attributes.