Add a data block for easier handling of references to an Alembic archive.
For now this ID-block only stores the filepath of the archive and it is
referenced by the Mesh Sequence Cache modifier and the Transform Cache
constraint.

The data block itself is called CacheFile for it to stay generic and be
reused for other file types other than Alembic files (similar to images
basically).