Curves: Auxiliary functions for the Extend operator
This commit contains part of the foundations for the Bezier curve Extend operator. This implementation is based on an addon created previously by a Blender user.

Some of the implemented functions may have use outside of the scope of the Extend operator. Therefore, they may be relocated and renamed in a future commit.