Add support to read and write multiple UV maps.
Whilst Alembic support only one UV layer per mesh, there is a convention
to add multiple UV layers as arbitrary geometry parameter for a given
mesh, so we use it to write all the UV layers to the Alembic files. Only
extra UVs in archives from external software which follow this
convention will be imported.