Cycles: Enable unaligned traversal for SSS and volume functions
Currently this traversal functions will always think there are unaligned
nodes in the scene and check visibility flags. This might not be fully
ideal, but on another hand adding more traversal functions might:

(a) Bump stack memory size on Maxwell GPUs (see the discussion on the
    Cycles mailing list)

(b) Not give real speed advantages.

So for now let's use simpler way and gather some benchmarks.