Wireframe/outline color from layer color-set
This commit makes wireframe/outline drawing of objects use the color-set of the highest layer in the hierarchy. This is an example of how a render engine (viewport in this case) can make use of the custom layer properties.
A local 3D View toggle allows quick toggling of the wireframe/outline colors.

On the general topic of custom wireframe colors: The layer based coloring makes it easy for users to assign colors that have meaning and are not just arbitrary choices. I discussed this with Ton, Sebastian König and others before, we all agreed that this is a pretty sane solution.
Also note that the colors are based on theme settings, which was a requirenment to avoid theme contrast issues.

Based on P158 by @campbellbarton.