Cycles: initial support for motion bluring of dynamic meshes.
This adds support for motion bluring of dynamic meshes (such as meshes
from liquid simulation).

For now, this is a little buggy since it appears that half of the motion
blur data is missing, also motion vectors seems to need to be scaled,
but I'm not sure just yet how to do so, and with which metric.