Always add object to active layer
I doubt all cases in which we have to assign an object after creation to a layer are covered, but this should be good enough for the start.
It's also quite easy to break object creation since there isn't any logic that ensures there's always an active layer (or any layer at all).