Added a roughness parameter for refractions (for scattering of the rays within an object)
With this, one can create a translucent material with a smooth surface and
with a milky look.

The final refraction roughness has to be calculated using the surface
roughness and the refraction roughness because those two are correlated
for refractions. If a ray hits a rough surface of a translucent material,
it is scattered while entering the surface. Then it is scattered further
within the object. The calculation I'm using is the following:

RefrRoughnessFinal = 1.0 - (1.0 - Roughness) * (1.0 - RefrRoughness)