Make move to layer OP work with new local view data
And fix crash on redo.

Current behavior is the same as in master, which could be improved. If in local view, the move to layer OP just opens a confirmation popup saying "Move out of Local View". Confirming only moves the object out of local view, you have to use the redo panel to move it to a different layer.