Curves: Offset operator
Fully implemented the modal operator for the offset tool. Also fixed a few bugs.There is still a bug when calculating the correct offset direction. If the distance is sufficiently high, or if the curve is too tight on a few places, the reverse_offset function returns true even though that is not what we want. Passing in a different BezTriple offset by a precision distance (such as 10e-5) should fix it, though attempts at this have proved unsuccessful.

Regarding editmesh_bevel.c, fixed a trailing whitespace that I found by chance.