Curves: Fillet and chamfer
Fixed a calculation error in the fillet operator. Limited the chamfer algorithm to just run on the points inbetween the first and the last handles, since it makes no sense to chamfer the start of a line. This brings a problem with cyclic splines though, which will be addressed in a later commit.