Curves: fillet and chamfer operator
The modal operators for the fillet and chamfer operators are now 90% complete.
On the chamfer operator, an option is needed to handle angle numinput. It would also be nice to have some sort of circular input for the angle (currently it is linked to the x variation of the cursor position).
For fillet, there is some kind of bug that corrupts curve data.

Due to the above mentioned reasons, modal operators are still commented.