Implement generic draw-modes and draw-plates for viewport engines
Also made the new viewport use these (at least a default mode with a number of default plates).

This draw-plate design is more generic then the previous hardcoded one, it stores the plates in mutable lists which can be stored as presets (draw-modes). We can further add sub-plates and process callbacks for plates that need additional data processing or don't do any drawing.