Cycles: Refactor so all split kernels have same signature
This is to set things up for supporting the split kernel on cpu and other
devices. By having the same signature for each kernel we can deduplicate
a lot of code without needing any trickery. The only kernel that doesn't
share this signature is the `data_init` kernel, which might end up being
different for each device type.