Cycles: Begin moving split kernel logic into own class
The new class `DeviceSplitKernel` will handle all logic for enqueueing of
kernels and memory memory management for the split kernel. Devices that
support the split kernel will create an instance of this class and call
its methods to run the split kernel.

There's still some work to do yet to make this device independent and to
deal with tile splitting.