Cycles: Remove tile splitting logic for split kernel
This is a huge reduction in code that was very hard to work with. It will be
much easier to go forward now that it is gone.

Unfortunate downside to not having tile splitting anymore is if the tile size
is set too large the user could get an out of memory error. We will have to
fix this at some point, but for now getting the kernel to run on CPU is
more important.

Also at this point the host side of the split kernel is almost entirely
independent of the underlying device, so work on CPU implementation can
finally start.