Cycles: Finish implementing the split kernel for CPU devices
Implements the `data_init` kernel plus a few fixes so that it runs correctly.

The split kernel now works on CPU devices. This means we have the ability to
debug the kernel, get better compiler logs, and have a new device to compare
to, all of which will help in developing and keeping the split kernel
correct. With this implementing branched path tracing, volumes and subsurface
scattering will be much easier.

Even at this point the CPU implementation already paying off by making long
existing bugs in the kernel apparent.

It should be noted that while the kernel works, it is a bit slower than the
mega kernel (~13% with BMW). Will need to investigate whether this is a
bug / implementation issue or if split kernels are simply slower by nature.