Make HMD window a proper 'restricted' window
A restricted window is a new concept that doesn't allow changing the screen of a window in any way. This way we can ensure the HMD window is only used for displaying the 3D View, so users can't accidentally change it without knowing how to get back. This also fixes some issues we had when using temp screens like they're used for UserPrefs.
Fixes:
* Could open toolbar/properties in HMD view
* HMD window used when opening UserPrefs through menu/shortcut
* Could easily change editor to non-3D-View